fix `chapter` & `read` table constraints
Turns out meddling with columns and constraints in SQLite requires
some... interesting gymnastics:

- turn foreign_keys pragma off
- rename existing table
- create new table with desired columns & constraints
- copy data over
- drop old table
- turn foreign_keys pragma back on

Also, you can't flip the foreign_keys pragma from inside a transaction,
so the whole migrator needed an update.

_Also_, even with foreign_keys=off, renaming `chapter` table to
`old_chapter` would automatically update the `read` table's foreign key
constraint to point to `old_chapter` too, which is NOT what we want. So
we need to set legacy_alter_table=on as well (as an extra bonus, this
behavior differs between sqlite versions as well - see the table at
https://www.sqlite.org/lang_altertable.html).